Made vote sorter not a service
